Works with the new architecture, need to do the doc and doxygen
