+++ applications_tests/lethe-particles/insert_clearing_with_file.prm
@@ -78,7 +78,7 @@ end
 #---------------------------------------------------
 
 subsection insertion info
-  set insertion method                = clear_and_add
+  set insertion method                = file
   set insertion frequency             = 100
   set list of input files             = ../particles_00.input, ../particles_01.input
   set clearing box points coordinates = 0., -1., -1.: 1., 1., 1.
