+++ applications_tests/lethe-particles/insert_clearing_with_filewith_dealii.geq.9.6.mpirun=1.output
