+++ include/dem/insertion_volume.h
@@ -41,10 +41,11 @@ public:
    * @param maximum_particle_diameter Maximum particle diameter based on values
    * defined in the parameter handler
    */
-  InsertionVolume(const DEMSolverParameters<dim> &dem_parameters,
-                  const double                    maximum_particle_diameter,
-                  const std::vector<std::shared_ptr<Distribution>>
-                    &distribution_object_container);
+  InsertionVolume(const std::vector<std::shared_ptr<Distribution>>
+                                                                  &distribution_object_container,
+                  const parallel::distributed::Triangulation<dim> &triangulation,
+                  const DEMSolverParameters<dim>                  &dem_parameters,
+                  const double                    maximum_particle_diameter);
 
   /**
    * Carries out the volume insertion of particles.
