+++ source/dem/dem.cc
@@ -8,7 +8,6 @@
 #include <dem/find_contact_detection_step.h>
 #include <dem/gear3_integrator.h>
 #include <dem/input_parameter_inspection.h>
-#include <dem/insertion_clear_and_add.h>
 #include <dem/insertion_file.h>
 #include <dem/insertion_list.h>
 #include <dem/insertion_plane.h>
@@ -950,40 +949,37 @@ std::shared_ptr<Insertion<dim>>
 DEMSolver<dim>::set_insertion_type(const DEMSolverParameters<dim> &parameters)
 {
   if (parameters.insertion_info.insertion_method ==
-      Parameters::Lagrangian::InsertionInfo::InsertionMethod::clear_and_add)
-    {
-      insertion_object = std::make_shared<InsertionClearAndAdd<dim>>(
-        parameters, triangulation, distribution_object_container);
-    }
-  else if (parameters.insertion_info.insertion_method ==
-           Parameters::Lagrangian::InsertionInfo::InsertionMethod::file)
+      Parameters::Lagrangian::InsertionInfo::InsertionMethod::file)
     {
       insertion_object =
-        std::make_shared<InsertionFile<dim>>(parameters,
-                                             distribution_object_container);
+        std::make_shared<InsertionFile<dim>>(distribution_object_container,
+                                             triangulation,
+                                             parameters);
     }
   else if (parameters.insertion_info.insertion_method ==
            Parameters::Lagrangian::InsertionInfo::InsertionMethod::list)
     {
       insertion_object =
-        std::make_shared<InsertionList<dim>>(parameters,
-                                             distribution_object_container);
+        std::make_shared<InsertionList<dim>>(distribution_object_container,
+                                             triangulation,
+                                             parameters);
     }
   else if (parameters.insertion_info.insertion_method ==
            Parameters::Lagrangian::InsertionInfo::InsertionMethod::plane)
     {
       insertion_object =
-        std::make_shared<InsertionPlane<dim>>(parameters,
+        std::make_shared<InsertionPlane<dim>>(distribution_object_container,
                                               triangulation,
-                                              distribution_object_container);
+                                              parameters);
     }
   else if (parameters.insertion_info.insertion_method ==
            Parameters::Lagrangian::InsertionInfo::InsertionMethod::volume)
     {
       insertion_object =
-        std::make_shared<InsertionVolume<dim>>(parameters,
-                                               maximum_particle_diameter,
-                                               distribution_object_container);
+        std::make_shared<InsertionVolume<dim>>(distribution_object_container,
+                                               triangulation,
+                                               parameters,
+                                               maximum_particle_diameter);
     }
   else
     {
